+++ wear/src/main/java/io/homeassistant/companion/android/onboarding/authentication/AuthenticationPresenter.kt
@@ -1,6 +1,6 @@
 package io.homeassistant.companion.android.onboarding.authentication
 
-interface PasswordAuthenticationPresenter {
+interface AuthenticationPresenter {
 
     fun onNextClicked(flowId: String, username: String?, password: String?, code: String?)
