--- wear/src/main/java/io/homeassistant/companion/android/onboarding/authentication/MfaAuthenticationPresenter.kt
@@ -1,8 +1,0 @@
-package io.homeassistant.companion.android.onboarding.authentication
-
-interface MfaAuthenticationPresenter {
-
-    fun onNextClicked(flowId: String, code: String)
-
-    fun onFinish()
-}
