--- wear/src/main/java/io/homeassistant/companion/android/onboarding/authentication/PasswordAuthenticationActivity.kt
@@ -1,91 +1,0 @@
-package io.homeassistant.companion.android.onboarding.authentication
-
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import android.util.Log
-import android.view.View
-import androidx.appcompat.app.AppCompatActivity
-import androidx.wear.activity.ConfirmationActivity
-import dagger.hilt.android.AndroidEntryPoint
-import io.homeassistant.companion.android.databinding.ActivityAuthenticationPasswordBinding
-import io.homeassistant.companion.android.onboarding.integration.MobileAppIntegrationActivity
-import javax.inject.Inject
-import io.homeassistant.companion.android.common.R as commonR
-
-@AndroidEntryPoint
-class PasswordAuthenticationActivity : AppCompatActivity(), PasswordAuthenticationView {
-    companion object {
-        private const val TAG = "PasswordAuthenticationActivity"
-
-        fun newInstance(context: Context, flowId: String): Intent {
-            var intent = Intent(context, PasswordAuthenticationActivity::class.java)
-            intent.putExtra("flowId", flowId)
-            return intent
-        }
-    }
-
-    @Inject
-    lateinit var presenter: PasswordAuthenticationPresenter
-    private lateinit var binding: ActivityAuthenticationPasswordBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        if (intent == null || !intent.hasExtra("flowId")) {
-            Log.e(TAG, "Flow id not specified, canceling authentication")
-            finish()
-        }
-
-        binding = ActivityAuthenticationPasswordBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.buttonNext.setOnClickListener {
-            presenter.onNextClicked(
-                intent.getStringExtra("flowId")!!,
-                binding.username.text.toString(),
-                binding.password.text.toString(),
-                binding.code.text.toString()
-            )
-        }
-    }
-
-    override fun onResume() {
-        super.onResume()
-
-        binding.loadingView.visibility = View.GONE
-    }
-
-    override fun startIntegration() {
-        startActivity(MobileAppIntegrationActivity.newInstance(this))
-    }
-
-    override fun showMfa() {
-        binding.titleLogin.visibility = View.GONE
-        binding.titleMfa.visibility = View.VISIBLE
-        binding.username.visibility = View.GONE
-        binding.password.visibility = View.GONE
-        binding.code.visibility = View.VISIBLE
-
-        binding.loadingView.visibility = View.GONE
-    }
-
-    override fun showLoading() {
-        binding.loadingView.visibility = View.VISIBLE
-    }
-
-    override fun showError() {
-        // Show failure message
-        val intent = Intent(this, ConfirmationActivity::class.java).apply {
-            putExtra(ConfirmationActivity.EXTRA_ANIMATION_TYPE, ConfirmationActivity.FAILURE_ANIMATION)
-            putExtra(ConfirmationActivity.EXTRA_MESSAGE, getString(commonR.string.failed_authentication))
-        }
-        startActivity(intent)
-        binding.loadingView.visibility = View.GONE
-    }
-
-    override fun onDestroy() {
-        presenter.onFinish()
-        super.onDestroy()
-    }
-}
