--- wear/src/main/java/io/homeassistant/companion/android/onboarding/authentication/PasswordAuthenticationModule.kt
@@ -1,14 +1,0 @@
-package io.homeassistant.companion.android.onboarding.authentication
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityComponent
-
-@Module
-@InstallIn(ActivityComponent::class)
-interface PasswordAuthenticationModule {
-
-    @Binds
-    fun authenticationPresenter(passwordAuthenticationPresenterImpl: PasswordAuthenticationPresenterImpl): PasswordAuthenticationPresenter
-}
