--- wear/src/main/java/io/homeassistant/companion/android/onboarding/authentication/PasswordAuthenticationView.kt
@@ -1,11 +1,0 @@
-package io.homeassistant.companion.android.onboarding.authentication
-
-interface PasswordAuthenticationView {
-    fun startIntegration()
-
-    fun showMfa()
-
-    fun showLoading()
-
-    fun showError()
-}
